app/requirements.txt
fastapi==0.119.0
uvicorn==0.22.0
python-dotenv==1.0.0
motor==3.7.1
pydantic==2.7.1